# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Code Intelligence Platform API",
    "description": "Comprehensive automation platform connecting OpenAI and GitHub APIs with advanced storage, real-time processing, and extensive third-party integrations",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-space-name.hf.space",
      "description": "Production server on Hugging Face Spaces"
    }
  ],
  "paths": {
...
    "/api/repositories/{id}/analyze": {
      "post": {
        "summary": "Start AI analysis of a repository",
        "operationId": "analyzeRepository",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "analysisType": {
                    "type": "string",
                    "enum": ["quality", "security", "performance", "documentation"]
                  },
                  "filePath": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Analysis started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflows": {
      "get": {
        "summary": "Get automation workflows",
        "operationId": "getWorkflows",
        "responses": {
          "200": {
            "description": "List of workflows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workflow"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new automation workflow",
        "operationId": "createWorkflow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "trigger": {
                    "type": "string",
                    "enum": ["push", "pull_request", "schedule", "manual"]
                  },
                  "config": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workflow"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Analysis": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "repositoryId": {
            "type": "integer"
          },
          "filePath": {
            "type": "string"
          },
          "analysisType": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "running", "completed", "failed"]
          },
          "qualityScore": {
            "type": "number"
          },
          "results": {
            "type": "object"
          }
        }
      },
...
      "Workflow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "lastRun": {
            "type": "string"
          }
        }
      },
...
    }
  }
}